refactor: rename handler v1 to v1beta1
