+++ api/handlers/v1beta1/grpc_rule_test.go
@@ -1,4 +1,4 @@
-package v1
+package v1beta1
 
 import (
 	"context"
