+++ api/handlers/v1beta1/grpc_subscription.go
@@ -1,4 +1,4 @@
-package v1
+package v1beta1
 
 import (
 	"context"
